with the schema update, solved the code breakings

+++ FundRaisingServer/Models/DTOs/Case/CaseResponseDto.cs
@@ -5,6 +5,6 @@ namespace FundRaisingServer.Models.DTOs
         public int CaseId { get; set; }
         public DateTime CreatedDate { get; set; }
         public double TotalDonations { get; set; }
-        public double CollectedDonations { get; set; }
+        public decimal CollectedDonations { get; set; }
     }
 }
